+++ ceammc/ext/src/math/mod_math.cpp
@@ -45,6 +45,7 @@ extern "C" void setup_math0x2eor();
 extern "C" void setup_math0x2eaapprox();
 
 void setup_math_round_tilde();
+void setup_math_cdiv_tilde();
 void setup_math_cmul_tilde();
 
 void ceammc_math_setup()
@@ -89,6 +90,7 @@ void ceammc_math_setup()
     setup_math0x2egcd();
     setup_math0x2elcm();
 
+    setup_math_cdiv_tilde();
     setup_math_cmul_tilde();
     setup_math_div();
     setup_math_expr();